+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 ### Unreleased
 - Fixed possible crash on some keys in `Symbol.for`
+- Fixed possible early error in `get-iterator-method` helper
 - Improved the way of inner iterators cleaning in iterator helpers
 - Improved accuracy of `Math.asinh` polyfill
 - Improved performance of `Uint8Array` base64 methods